refactor(core): relocate Automations cluster into app/automation.rs
ADR-22 step 2: move the Automations-cluster impl App methods out of
src/app/mod.rs and src/app/key_handlers.rs into a new src/app/automation.rs.
Mechanical relocation only — no behavior change, no field moves (state stays
in AutomationUiState as App::automation_ui). Cross-module moves widen the
relocated private methods to pub(crate); handle_task_editor_pane_key (stays
in key_handlers, called by the moved handle_automation_pane_capture) is also
bumped to pub(crate), and view.rs's format_automation_summary call now goes
through super::automation::.
