+++ src/app/key_handlers.rs
@@ -388,40 +388,6 @@ impl App {
         }
     }
 
-    /// The in-pane automation/task editor + run-history capture input like the
-    /// overlay modal — so editor chords (e.g. `e`, `d`, Ctrl+E) reach the form
-    /// instead of firing a global binding. Focus navigation (Ctrl+L/H) and quit
-    /// still pass through to the global handler so you can move between panes.
-    /// Returns `true` if consumed.
-    fn handle_automation_pane_capture(&mut self, code: KeyCode, mods: KeyModifiers) -> bool {
-        if !matches!(
-            self.focus,
-            InputFocus::AutomationEditor
-                | InputFocus::AutomationRunHistory
-                | InputFocus::TaskEditor
-        ) {
-            return false;
-        }
-        let passthrough = matches!(
-            self.keybindings.lookup(code, mods),
-            Some(
-                crate::session::Action::FocusForward
-                    | crate::session::Action::FocusBackward
-                    | crate::session::Action::QuitApp
-            )
-        );
-        if passthrough {
-            return false;
-        }
-        match self.focus {
-            InputFocus::AutomationEditor => self.handle_automation_editor_pane_key(code, mods),
-            InputFocus::AutomationRunHistory => self.handle_automation_run_history_key(code),
-            InputFocus::TaskEditor => self.handle_task_editor_pane_key(code, mods),
-            _ => unreachable!(),
-        }
-        true
-    }
-
     /// The ordered focus ring for the **current context**. `Ctrl+L`/`Ctrl+H`
     /// cycle *within* this ring; switching between the session and automation
     /// contexts is done with `j`/`k` in the left column (not the focus cycle).
@@ -504,170 +470,11 @@ impl App {
         self.refresh_automation_view();
     }
 
-    /// Handle keys while the automations pane is focused: navigate and drive the
-    /// same toggle/run/edit/delete actions as the Ctrl+P modal. (Ctrl+N to
-    /// create is handled globally in `dispatch_action`, so it works here too,
-    /// including on an empty pane.)
-    pub(crate) fn handle_automations_pane_key(&mut self, code: KeyCode) {
-        let count = self.automation_ui.cached_automations.len();
-        match code {
-            // Creating works regardless of whether the pane has entries.
-            KeyCode::Char('n') => self.new_automation_in_pane(),
-            // `k`/Up at the top row (or empty pane) flows focus back up into the
-            // session list; `j`/Down past the last loops to the top of it — so
-            // the left column behaves as one circular vertical list.
-            KeyCode::Char('k') | KeyCode::Up => self.automations_pane_move_up(count),
-            KeyCode::Char('j') | KeyCode::Down => self.automations_pane_move_down(count),
-            // `Enter`/`e` focuses the central-pane editor (like `Enter` on a
-            // session focuses its terminal); on an empty pane it starts a new
-            // automation.
-            KeyCode::Enter | KeyCode::Char('e') => self.enter_automation_editor_in_pane(count),
-            // Remaining nav/actions are no-ops on an empty pane.
-            _ if count > 0 => self.dispatch_automation_pane_action(code, count),
-            _ => {}
-        }
-    }
-
-    /// `Enter`/`e` in the automations pane: open the editor for the selection,
-    /// or start a new automation on an empty pane.
-    fn enter_automation_editor_in_pane(&mut self, count: usize) {
-        if count == 0 {
-            self.new_automation_in_pane();
-        } else {
-            self.enter_automation_editor();
-        }
-        self.refresh_selected_automation_runs();
-    }
-
-    /// Clamp the selection and run the toggle/run/delete action for the row
-    /// under the cursor (caller guarantees a non-empty pane).
-    fn dispatch_automation_pane_action(&mut self, code: KeyCode, count: usize) {
-        if self.automation_ui.automation_panel_index >= count {
-            self.automation_ui.automation_panel_index = count - 1;
-        }
-        let id =
-            self.automation_ui.cached_automations[self.automation_ui.automation_panel_index].id;
-        self.handle_automation_pane_action(code, id);
-    }
-
-    /// `k`/Up in the automations pane: step up, or hand focus back to the
-    /// session list (last row) at the top / when empty.
-    fn automations_pane_move_up(&mut self, count: usize) {
-        if self.automation_ui.automation_panel_index == 0 || count == 0 {
-            self.focus = InputFocus::SessionList;
-            self.select_last_session();
-        } else {
-            self.automation_ui.automation_panel_index -= 1;
-        }
-        self.refresh_automation_view();
-    }
-
-    /// `j`/Down in the automations pane: step down, or loop focus to the top of
-    /// the session list past the last row / when empty.
-    fn automations_pane_move_down(&mut self, count: usize) {
-        if count == 0 || self.automation_ui.automation_panel_index + 1 >= count {
-            self.focus = InputFocus::SessionList;
-            self.select_first_session();
-        } else {
-            self.automation_ui.automation_panel_index += 1;
-        }
-        self.refresh_automation_view();
-    }
-
-    /// Toggle / run / delete the selected automation (`Space`/`r`/`d`).
-    fn handle_automation_pane_action(&mut self, code: KeyCode, id: i64) {
-        match code {
-            KeyCode::Char(' ') => {
-                self.toggle_automation_by_id(id);
-                self.sync_automation_editor();
-            }
-            KeyCode::Char('r') => self.run_automation_by_id(id),
-            KeyCode::Char('d') => {
-                self.delete_automation_by_id(id);
-                let new_count = self.automation_ui.cached_automations.len();
-                if new_count > 0 && self.automation_ui.automation_panel_index >= new_count {
-                    self.automation_ui.automation_panel_index = new_count - 1;
-                }
-                self.refresh_automation_view();
-            }
-            _ => {}
-        }
-    }
-
-    /// Handle keys while editing the scoped automation in the central pane.
-    /// `Enter` saves (and returns to the list), `Esc` discards; field navigation
-    /// is shared with the overlay editor. `Ctrl+L`/`Ctrl+H` are handled earlier
-    /// as global focus actions, so they move focus out of / back into the editor.
-    fn handle_automation_editor_pane_key(&mut self, code: KeyCode, mods: KeyModifiers) {
-        // No editor yet (e.g. focused an empty pane): allow create / leave only.
-        let Some(editor) = self.automation_ui.automation_editor.as_mut() else {
-            match code {
-                KeyCode::Char('n') => self.new_automation_in_pane(),
-                KeyCode::Esc => self.focus = InputFocus::Automations,
-                _ => {}
-            }
-            return;
-        };
-        match editor.handle_key(code, mods) {
-            super::modals::EditorOutcome::Continue => {}
-            super::modals::EditorOutcome::Save => {
-                let Some(editor) = self.automation_ui.automation_editor.clone() else {
-                    return;
-                };
-                if self.save_automation(&editor) {
-                    // A brand-new automation lands at the top of the list.
-                    if editor.editing_id.is_none() {
-                        self.automation_ui.automation_panel_index = 0;
-                    }
-                    self.focus = InputFocus::Automations;
-                    self.refresh_automation_view();
-                }
-            }
-            super::modals::EditorOutcome::Cancel => {
-                // Discard edits and restore the preview for the selection.
-                self.focus = InputFocus::Automations;
-                self.refresh_automation_view();
-            }
-        }
-    }
-
-    /// Handle keys while the run-history panel is focused: `j`/`k` move the
-    /// selected run, `r` triggers a fresh run of the scoped automation, and
-    /// `Esc` returns to the editor.
-    fn handle_automation_run_history_key(&mut self, code: KeyCode) {
-        match code {
-            KeyCode::Char('j') | KeyCode::Down => {
-                let len = self.automation_ui.cached_automation_runs.len();
-                if len > 0 && self.automation_ui.automation_run_index + 1 < len {
-                    self.automation_ui.automation_run_index += 1;
-                }
-            }
-            KeyCode::Char('k') | KeyCode::Up => {
-                self.automation_ui.automation_run_index =
-                    self.automation_ui.automation_run_index.saturating_sub(1);
-            }
-            // Trigger a fresh run of the scoped automation now. (`run_automation_by_id`
-            // refreshes the caches; the new run record lands on a later tick.)
-            KeyCode::Char('r') => {
-                if let Some(id) = self.scoped_automation_id() {
-                    self.run_automation_by_id(id);
-                }
-            }
-            // Jump to the session this run touched (the send target / spawned
-            // session), when it's still open.
-            KeyCode::Enter => {
-                self.open_run_related_session();
-            }
-            KeyCode::Esc => self.focus = InputFocus::AutomationEditor,
-            _ => {}
-        }
-    }
-
     /// Handle keys while editing the scoped task in the central pane. `Enter`
     /// saves and returns to the tasks panel; `Esc` discards and returns. Field
     /// navigation is the `TaskEditorModal`'s own. Mirrors
     /// `handle_automation_editor_pane_key`.
-    fn handle_task_editor_pane_key(&mut self, code: KeyCode, mods: KeyModifiers) {
+    pub(crate) fn handle_task_editor_pane_key(&mut self, code: KeyCode, mods: KeyModifiers) {
         let Some(editor) = self.task_ui.task_editor.as_mut() else {
             match code {
                 KeyCode::Char('n') => self.new_task_in_pane(),
@@ -994,105 +801,6 @@ impl App {
             } else {
                 // Normal flow — proceed to role selection / spawn.
                 self.finish_prepare_spawn(name, config, worktrees);
-            }
-        }
-    }
-
-    /// Drive the centered-overlay automation editor (the Ctrl+P list path).
-    /// Field navigation is shared with the in-pane editor via
-    /// [`AutomationEditorModal::handle_key`].
-    fn handle_automation_editor_key(&mut self, code: KeyCode, mods: KeyModifiers) {
-        let super::modals::Modal::AutomationEditor(ref mut m) = self.modal else {
-            return;
-        };
-        match m.handle_key(code, mods) {
-            super::modals::EditorOutcome::Continue => {}
-            super::modals::EditorOutcome::Save => self.submit_automation_editor(),
-            super::modals::EditorOutcome::Cancel => self.modal.close(),
-        }
-    }
-
-    fn handle_automations_list_key(&mut self, code: KeyCode) {
-        if self.handle_automations_list_nav(code) {
-            return;
-        }
-        self.handle_automations_list_action(code);
-    }
-
-    /// Close/select navigation for the automations list modal. Returns `true`
-    /// when the key was consumed as navigation.
-    fn handle_automations_list_nav(&mut self, code: KeyCode) -> bool {
-        let super::modals::Modal::AutomationsList(ref mut al) = self.modal else {
-            return false;
-        };
-        match code {
-            KeyCode::Esc => self.modal.close(),
-            KeyCode::Char('j') | KeyCode::Down => {
-                if al.index + 1 < al.entries.len() {
-                    al.index += 1;
-                }
-            }
-            KeyCode::Char('k') | KeyCode::Up => {
-                al.index = al.index.saturating_sub(1);
-            }
-            _ => return false,
-        }
-        true
-    }
-
-    /// Action keys (new/edit/toggle/run/delete) for the automations list modal.
-    fn handle_automations_list_action(&mut self, code: KeyCode) {
-        match code {
-            KeyCode::Char('n') => {
-                self.modal.close();
-                self.open_automation_editor();
-            }
-            KeyCode::Char('e') | KeyCode::Enter => {
-                if let Some(id) = self.selected_automation_id() {
-                    self.modal.close();
-                    self.open_edit_automation(id);
-                }
-            }
-            KeyCode::Char(' ') => {
-                if let Some(id) = self.selected_automation_id() {
-                    self.toggle_automation_by_id(id);
-                    self.refresh_automations_list_modal();
-                }
-            }
-            KeyCode::Char('r') => {
-                if let Some(id) = self.selected_automation_id() {
-                    self.run_automation_by_id(id);
-                }
-            }
-            KeyCode::Char('d') => {
-                if let Some(id) = self.selected_automation_id() {
-                    self.delete_automation_by_id(id);
-                    self.refresh_automations_list_modal();
-                }
-            }
-            _ => {}
-        }
-    }
-
-    /// The id of the selected automation in the list modal, if any.
-    fn selected_automation_id(&self) -> Option<i64> {
-        let super::modals::Modal::AutomationsList(ref al) = self.modal else {
-            return None;
-        };
-        al.entries.get(al.index).map(|e| e.id)
-    }
-
-    /// Rebuild the list modal entries after a mutation, preserving selection.
-    fn refresh_automations_list_modal(&mut self) {
-        let index = match self.modal {
-            super::modals::Modal::AutomationsList(ref al) => al.index,
-            _ => return,
-        };
-        self.open_automations_list();
-        if let super::modals::Modal::AutomationsList(ref mut al) = self.modal {
-            al.index = index.min(al.entries.len().saturating_sub(1));
-            if al.entries.is_empty() {
-                self.modal.close();
             }
         }
     }
